requirements.txt
numpy>=1.21.0
SimpleITK>=2.1.0
pandas>=1.3.0
click>=8.0.0
tqdm>=4.62.0
matplotlib>=3.4.0
scipy>=1.7.0
scikit-image>=0.18.0
pyradiomics>=3.0.0
napari>=0.4.12
PyYAML>=6.0.0
torch>=1.9.0  # Optional for deep learning features
